requirements.txt
pyyaml
paramiko
pillow
numpy
spydev
rpi_ws281x